finish dev command refactor
